Cancelling participations

+++ src/main/java/pl/sii/itconference/dto/CreateParticipationDto.java
@@ -18,5 +18,5 @@ public class CreateParticipationDto {
     private String email;
 
     @NotNull
-    private Integer lectureId;
+    private Long lectureId;
 }
